requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
backtrader==1.9.76.123
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.9
cycler==0.10.0
Django==3.1.7
gunicorn==20.0.4
idna==2.10
idna-ssl==1.1.0
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.4
mplfinance==0.12.7a5
multidict==5.1.0
numpy==1.19.5
pandas==1.1.5
pandas-datareader==0.9.0
Pillow==8.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywinauto==0.6.6
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
slackclient==2.9.3
slacker==0.14.0
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
